Delegation is overkill, stop that
